Add Modal launcher, derived schedule defaults, and bfloat16 RoPE fix
The training script now defaults the peak learning rate to 1e-3 and
derives warmup, cosine length, and the final learning rate from
train-iters and lrmax unless passed explicitly, and prints the token
budget at startup. RoPE casts its cos and sin tables to the input dtype
so bfloat16 runs no longer mix dtypes in attention. modal_train.py runs
single jobs or parallel learning-rate sweeps on Modal GPUs with data
and run outputs on volumes.

+++ cs336_basics/transformer.py
@@ -90,8 +90,8 @@ class RotaryPositionalEmbedding(nn.Module):
 
     def forward(self, x, token_positions):
         x_rearr = rearrange(x, "... seq_len (pair_d_k two) -> ... seq_len pair_d_k two", two=2)
-        cos = self.Rcos[token_positions]
-        sin = self.Rsin[token_positions]
+        cos = self.Rcos[token_positions].to(x.dtype)
+        sin = self.Rsin[token_positions].to(x.dtype)
 
         out_even = cos * x_rearr[..., 0] - sin * x_rearr[..., 1]
         out_odd = sin * x_rearr[..., 0] + cos* x_rearr[..., 1]
